externalize database helper
